refactor: rewrite gen-esbuild-test in TypeScript (#7376)
Rewrote the gen-esbuild-test script in TypeScript with AI.
